Refactored ComposeException to apiserver from registration and policy services
