+++ src/main/java/iudx/aaa/server/policy/CatalogueClient.java
@@ -13,6 +13,7 @@ import io.vertx.sqlclient.SqlResult;
 import io.vertx.sqlclient.Tuple;
 import iudx.aaa.server.apiserver.ResourceObj;
 import iudx.aaa.server.apiserver.Response;
+import iudx.aaa.server.apiserver.util.ComposeException;
 import iudx.aaa.server.apiserver.util.Urn;
 import org.apache.log4j.LogManager;
 import org.apache.log4j.Logger;
